requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
colorlog==4.2.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
distlib==0.3.1
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
filelock==3.0.12
har2case==0.3.1
HttpRunner==1.5.8
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.4
pipenv==2020.6.2
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.27
virtualenv-clone==0.5.4
zipp==3.1.0
